refactor: 일정 삭제 service 로직 책임 분리 및 테스트 코드 리팩토링
